refactor: Listener 롤백
